+++ backend/packages/caching/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @cat-factory/caching
 
+## 0.12.12
+
+### Patch Changes
+
+- @cat-factory/kernel@0.214.1
+
+## 0.12.11
+
+### Patch Changes
+
+- Updated dependencies [3435bd1]
+  - @cat-factory/kernel@0.214.0
+
 ## 0.12.10
 
 ### Patch Changes