+++ backend/packages/provider-s3/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @cat-factory/provider-s3
 
+## 0.2.270
+
+### Patch Changes
+
+- @cat-factory/kernel@0.214.1
+
+## 0.2.269
+
+### Patch Changes
+
+- Updated dependencies [3435bd1]
+  - @cat-factory/kernel@0.214.0
+
 ## 0.2.268
 
 ### Patch Changes